backend/requirements.txt
flask==2.3.2
pyserial==3.5
